[Project] Further split of the code
