+++ src/libserver/CMakeLists.txt
@@ -21,6 +21,8 @@ SET(LIBRSPAMDSERVERSRC
 				${CMAKE_CURRENT_SOURCE_DIR}/spf.c
 				${CMAKE_CURRENT_SOURCE_DIR}/ssl_util.c
 				${CMAKE_CURRENT_SOURCE_DIR}/symcache/symcache_impl.cxx
+				${CMAKE_CURRENT_SOURCE_DIR}/symcache/symcache_item.cxx
+				${CMAKE_CURRENT_SOURCE_DIR}/symcache/symcache_runtime.cxx
 				${CMAKE_CURRENT_SOURCE_DIR}/symcache/symcache_c.cxx
 				${CMAKE_CURRENT_SOURCE_DIR}/task.c
 				${CMAKE_CURRENT_SOURCE_DIR}/url.c
